+++ lib/state_pulse.dart
@@ -3,3 +3,6 @@
 export 'src/hydrated_state_pulse.dart';
 export 'src/state_pulse_provider.dart';
 export 'src/state_pulse_builder.dart';
+export 'src/state_pulse_selector.dart';
+export 'src/state_pulse_listener.dart';
+export 'src/state_pulse_consumer.dart';